@@ -11,33 +11,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.0.3",
-    "@angular/common": "^13.0.3",
-    "@angular/compiler": "^13.0.3",
-    "@angular/core": "^13.0.3",
-    "@angular/forms": "^13.0.3",
-    "@angular/localize": "^13.0.3",
-    "@angular/platform-browser": "^13.0.3",
-    "@angular/platform-browser-dynamic": "^13.0.3",
-    "@angular/router": "^13.0.3",
-    "@angular/cdk": "^13.0.4",
-    "@swimlane/ngx-charts": "^19.1.0",
-    "@swimlane/ngx-datatable": "^20.0.0",
-    "bootstrap": "^5.1.3",
+    "@angular/animations": "^16.2.4",
+    "@angular/common": "^16.2.4",
+    "@angular/compiler": "^16.2.4",
+    "@angular/core": "^16.2.4",
+    "@angular/forms": "^16.2.4",
+    "@angular/localize": "^16.2.4",
+    "@angular/platform-browser": "^16.2.4",
+    "@angular/platform-browser-dynamic": "^16.2.4",
+    "@angular/router": "^16.2.4",
+    "@angular/cdk": "^16.2.3",
+    "@swimlane/ngx-charts": "^20.4.1",
+    "@swimlane/ngx-datatable": "^20.1.0",
+    "bootstrap": "^5.3.1",
     "chokidar": "^3.5.2",
     "jquery": "^3.6.0",
-    "ngx-cookie-service": "^13.0.1",
-    "ngx-filter-pipe": "^2.1.2",
+    "ngx-cookie-service": "^16.0.1",
     "@popperjs/core": "^2.11.0",
     "rxjs": "^6.6.3",
     "tslib": "^2.3.1",
-    "zone.js": "^0.11.4"
+    "zone.js": "^0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.0.4",
-    "@angular/cli": "^13.0.4",
-    "@angular/compiler-cli": "^13.0.3",
-    "@angular/language-service": "^13.0.3",
+    "@angular-devkit/build-angular": "^16.2.1",
+    "@angular/cli": "^16.2.1",
+    "@angular/compiler-cli": "^16.2.4",
+    "@angular/language-service": "^16.2.4",
     "@types/jasmine": "^3.10.2",
     "@types/jasminewd2": "^2.0.10",
     "@types/node": "^16.11.12",
@@ -48,9 +47,7 @@
     "karma-coverage-istanbul-reporter": "^3.0.3",
     "karma-jasmine": "^4.0.1",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "protractor": "^7.0.0",
-    "ts-node": "^10.4.0",
     "eslint": "^7.32.0",
-    "typescript": "~4.4.3"
+    "typescript": "~5.1.6"
   }
 }
